# api/swagger/swagger.yaml
swagger: '2.0'
info: {version: 1.0.0, title: Users-App, description: APIs exposing RESTful Users
    App endpoints}
host: localhost:3000
basePath: /api
schemes: [http]
consumes: [application/json, multipart/form-data]
produces: [application/json]
definitions:
  v0@Response:
    type: object
    properties:
      data:
        type: object
        properties:
          first_name: {type: string}
          id: {type: string}
  v0@ErrorResponse:
    required: [message]
    properties:
      message: {type: string}
paths:
...
  /v0/search_users:
    x-swagger-router-controller: v0/components/users/controller
    get:
      tags: [v0 | User]
      summary: Search users list
      description: get the users list
      operationId: searchUsers
      parameters:
      - {name: authorization, in: header, required: true, type: string, description: Auth
          token}
      - {name: name, in: query, type: string, description: name}
      - {name: contact, in: query, type: string, description: contact}
      responses:
        '200':
          description: Success
          schema: {$ref: '#/definitions/v0@Response'}
        '204': {description: No content}
        default:
          description: Error
          schema: {$ref: '#/definitions/v0@ErrorResponse'}
...